New Camera Animations, Sprint conditions, Dodge conditions
+ New script: CamAnimation.cs
+ New player camera animations that plays depending on states: walking, sprinting, dodging.

* Added an additional condition for sprint:
Only 'W' key must be held to activate and maintain sprinting state.

* Added an additional condition and edited code for dodge:
Player can now only dodge towards left or right, 'A' or 'D' key MUST now be held to activate dodge, if 'W' or 'S' is held as well, the forward/backward value added to the dodge direction is halved.

* Changed stamina regen rate and amount in the scene to tick faster with a smaller amount.

* Lowered the stamina consumption rate of sprinting from 0.5 to 0.05. Reason: sprinting depleted the stamina too fast, making sprinting strategically very inefficient.